requirements.txt
python-dotenv==0.19.0
beautifulsoup4==4.9.3
selenium==3.141.0
spacy==2.2.3
scikit-learn==0.24.2
pandas==1.3.2
sqlalchemy==1.4.23
fastapi==0.68
fastapi_utils==0.2.1
pydantic==1.7.4
requests==2.*
uvicorn[standard]==0.15.0

requests